+++ pika/third/blackwidow/src/redis_strings.cc
@@ -33,6 +33,8 @@ Status RedisStrings::Open(const BlackwidowOptions& bw_options,
   ops.table_factory.reset(rocksdb::NewBlockBasedTableFactory(table_ops));
 
   ops.min_blob_size = bw_options.min_blob_size;
+  ops.rate_limiter = bw_options.rate_limiter;
+  default_write_options_.disableWAL = bw_options.disable_wal;
 
   return rocksdb::titandb::TitanDB::Open(ops, db_path, &Titandb_);
 }
@@ -63,12 +65,11 @@ Status RedisStrings::ScanKeyNum(uint64_t* num) {
 
   // Note: This is a string type and does not need to pass the column family as
   // a parameter, use the default column family
-  rocksdb::Iterator* iter = Titandb_->NewIterator(iterator_options);
+  rocksdb::Iterator* iter = Titandb_->NewKeyIterator(iterator_options);
   for (iter->SeekToFirst();
        iter->Valid();
        iter->Next()) {
-    ParsedStringsValue parsed_strings_value(iter->value());
-    if (!parsed_strings_value.IsStale()) {
+    if (!iter->IsStale()) {
       count++;
     }
   }
@@ -88,12 +89,11 @@ Status RedisStrings::ScanKeys(const std::string& pattern,
 
   // Note: This is a string type and does not need to pass the column family as
   // a parameter, use the default column family
-  rocksdb::Iterator* iter = Titandb_->NewIterator(iterator_options);
+  rocksdb::Iterator* iter = Titandb_->NewKeyIterator(iterator_options);
   for (iter->SeekToFirst();
        iter->Valid();
        iter->Next()) {
-    ParsedStringsValue parsed_strings_value(iter->value());
-    if (!parsed_strings_value.IsStale()) {
+    if (!iter->IsStale()) {
       key = iter->key().ToString();
       if (StringMatch(pattern.data(),
             pattern.size(), key.data(), key.size(), 0)) {
@@ -101,6 +101,7 @@ Status RedisStrings::ScanKeys(const std::string& pattern,
       }
     }
   }
+
   delete iter;
   return Status::OK();
 }
@@ -306,9 +307,10 @@ Status RedisStrings::Decrby(const Slice& key, int64_t value, int64_t* ret) {
       StringsValue strings_value(new_value);
       return Titandb_->Put(default_write_options_, key, strings_value.Encode());
     } else {
-      parsed_strings_value.StripSuffix();
+      int32_t timestamp = parsed_strings_value.timestamp();
+      std::string old_user_value = parsed_strings_value.value().ToString();
       char* end = nullptr;
-      int64_t ival = strtoll(old_value.c_str(), &end, 10);
+      int64_t ival = strtoll(old_user_value.c_str(), &end, 10);
       if (*end != 0) {
         return Status::Corruption("Value is not a integer");
       }
@@ -319,6 +321,7 @@ Status RedisStrings::Decrby(const Slice& key, int64_t value, int64_t* ret) {
       *ret = ival - value;
       new_value = std::to_string(*ret);
       StringsValue strings_value(new_value);
+      strings_value.set_timestamp(timestamp);
       return Titandb_->Put(default_write_options_, key, strings_value.Encode());
     }
   } else if (s.IsNotFound()) {
@@ -521,9 +524,10 @@ Status RedisStrings::Incrby(const Slice& key, int64_t value, int64_t* ret) {
       StringsValue strings_value(buf);
       return Titandb_->Put(default_write_options_, key, strings_value.Encode());
     } else {
-      parsed_strings_value.StripSuffix();
+      int32_t timestamp = parsed_strings_value.timestamp();
+      std::string old_user_value = parsed_strings_value.value().ToString();
       char* end = nullptr;
-      int64_t ival = strtoll(old_value.c_str(), &end, 10);
+      int64_t ival = strtoll(old_user_value.c_str(), &end, 10);
       if (*end != 0) {
         return Status::Corruption("Value is not a integer");
       }
@@ -535,6 +539,7 @@ Status RedisStrings::Incrby(const Slice& key, int64_t value, int64_t* ret) {
       char buf[32];
       Int64ToStr(buf, 32, *ret);
       StringsValue strings_value(buf);
+      strings_value.set_timestamp(timestamp);
       return Titandb_->Put(default_write_options_, key, strings_value.Encode());
     }
   } else if (s.IsNotFound()) {
@@ -565,10 +570,11 @@ Status RedisStrings::Incrbyfloat(const Slice& key, const Slice& value,
       StringsValue strings_value(new_value);
       return Titandb_->Put(default_write_options_, key, strings_value.Encode());
     } else {
-      parsed_strings_value.StripSuffix();
+      int32_t timestamp = parsed_strings_value.timestamp();
+      std::string old_user_value = parsed_strings_value.value().ToString();
       long double total, old_number;
-      if (StrToLongDouble(old_value.data(),
-                          old_value.size(), &old_number) == -1) {
+      if (StrToLongDouble(old_user_value.data(),
+                          old_user_value.size(), &old_number) == -1) {
         return Status::Corruption("Value is not a vaild float");
       }
       total = old_number + long_double_by;
@@ -577,6 +583,7 @@ Status RedisStrings::Incrbyfloat(const Slice& key, const Slice& value,
       }
       *ret = new_value;
       StringsValue strings_value(new_value);
+      strings_value.set_timestamp(timestamp);
       return Titandb_->Put(default_write_options_, key, strings_value.Encode());
     }
   } else if (s.IsNotFound()) {
@@ -1337,27 +1344,39 @@ Status RedisStrings::Persist(const Slice& key) {
 }
 
 Status RedisStrings::TTL(const Slice& key, int64_t* timestamp) {
-  std::string value;
-  ScopeRecordLock l(lock_mgr_, key);
-  Status s = Titandb_->Get(default_read_options_, key, &value);
+  int32_t ts;
+  Status s = Titandb_->GetTimestamp(default_read_options_, key, &ts);
   if (s.ok()) {
-    ParsedStringsValue parsed_strings_value(&value);
-    if (parsed_strings_value.IsStale()) {
-      *timestamp = -2;
-      return Status::NotFound("Stale");
+    if (0 == ts) {
+      *timestamp = -1;
     } else {
-      *timestamp = parsed_strings_value.timestamp();
-      if (*timestamp == 0) {
-        *timestamp = -1;
+      int64_t unix_time;
+      rocksdb::Env::Default()->GetCurrentTime(&unix_time);
+      if (ts < unix_time) {
+        *timestamp = -2;
+        return Status::NotFound("Stale");
       } else {
-        int64_t curtime;
-        rocksdb::Env::Default()->GetCurrentTime(&curtime);
-        *timestamp = *timestamp - curtime >= 0 ? *timestamp - curtime : -2;
+        *timestamp = ts - unix_time >= 0 ? ts - unix_time : -2;
       }
-    }
+    } 
   } else if (s.IsNotFound()) {
     *timestamp = -2;
   }
+
+  return s;
+}
+
+Status RedisStrings::Exists(const Slice& key) {
+  int32_t ts;
+  Status s = Titandb_->GetTimestamp(default_read_options_, key, &ts);
+  if (s.ok() && (0 != ts)) {
+    int64_t unix_time;
+    rocksdb::Env::Default()->GetCurrentTime(&unix_time);
+    if (ts < unix_time) {
+      return Status::NotFound("Stale");
+    }
+  }
+
   return s;
 }